Use a MessageInput sequenceNr instead of journalOffset
Introduce a seqence number on each input which gets incremented
and embedded in every received message.
This allows sorting to work without depending on a KafkaJournal.

Regenerate the protobuf class JournalMessages so it can pass the
sequence number from a RawMessage to a Message.

Don't overwrite already existing GL2_MESSAGE_IDs.
They might already be set if we are receiving messages from a Graylog forwarder.