Prevent stale objects in stream cache of route_to_stream function (#6788)
Previously we were storing the stream itself in the nameToStream
MultiMap but were using only the id of the stream as the comparator for
the underlying set. This would lead to stale stream objects in that
cache if the cache was updated with a changed stream which was already
cached.

The implementation felt way to complicated for what it was doing. For
the benefit of having clearer semantics we simply reload all streams
when a stream changes. This will put additional load on the database for
cache updating but cache lookup times will stay stable.

Fixes #4954