Include `messages` field if empty in message list search type result. (#6872)
In 5ee5115642dd0c8d6a1f7ed58f68317a9ccda3f0 an optional name attribute
defined by the caller for a search type, which is passed through to its
results, was introduced. To leave it out when it is empty, the
`@JsonInclude(NON_EMPTY)` annotation was added. Unfortunately, this also
leaves out the `messages` in the result if it is empty, breaking
consumer logic which expects it to be present.

This change turns it into `NON_ABSENT` which skips `Optional`s which are
empty, but includes empty collections.