Improve error propagation
When the URL is not whitelisted, we want that error to be exposed so
that it can be displayed on the frontend. When the url becomes
whitelisted, we want the error to go away and any previous or new errors
related to parsing the DSV file or requesting the file should be
exposed.

To make this work reliably, we have to make sure that in case of a
whitelist error the file is fetched again so that the errors related
to file fetching and/or processing surface again. Also we want errors
related to parsing the file show up in the UI, which was not the case
before when the file became corrupted at the source when the data
adapter was already running. Those errors would only show up when the
data adapter was (re-)started.