Implement error boundary for widgets.
Prior to this change, a malfunctioning widget implementation (throwing
an error intentionally or accidentally) would lead to the general error
page and the search/dashboard would not be usable.

This change implements an error boundary in the `Widget` component, that
catches all errors thrown by components below it in the component tree
and displays them. This allows the user to operate on the
search/dashboard to fix the issue and continues showing other, working
widgets.