Add testcontainers based MongoDBInstance rule (#6880)
* Add testcontainers based MongoDBInstance rule

This rule can be used for integration tests that need a MongoDB
database. It is using the testcontainers library to start a database
instance.

It supports our standard database fixture files by using the
MongoDBFixtures annotation on test methods.

* Remove MongoConnection#getMongoClient()

This was added for a previous iteration of the MongoDBInstance change
but is not needed anymore.