Fix "COMMONAPACHELOG" grok pattern. (#6847)
* Fix COMMONAPACHELOG grok pattern.

The COMMONAPACHELOG grok pattern uses the original date string found in
the source to propagate the "timestamp" message field. When the
extractor tries to use this value to set the field on the message, it
will be ignored, as it is not a proper date.

We adjust the pattern so that it creates a date object which the message
accepts as a valid value for the "timestamp" field.

Fixes #1647

* Add migration to fix default grok patterns

Adds a migration which will fix a grok pattern once, if it is still in
a state that needs fixing.