Fall back to first search type result to extract effective timerange. (#6943)
In 81107aacbc827a4c2d3d207685b4b544541338cb we introduced the capability
to set names for search types which are piped through to the results,
allowing a widget to generate multiple search types and address the
results by name. Searches created before this, result in the search type
id being used instead of the name when generating the result object.

Prior to this PR, extracting the effective timerange from the result
object incorrectly assumed that it has a `chart` key. This does not hold
true for aforementioned searches created before the introduction of the
`name` property. This introduced a regression, where a widget relying on
the effective timerange failed.

This PR is addressing this, by falling back to the first search type
result from the result object for the extraction of the effective
timerange.

Fixes #6944