Supporting widget override elements.
Widget override elements can be registered and are rendered when editing
a widget. These component can either return elements which are rendered
next to the widget, or they can throw a renderable component which will
be rendered instead of the component.

The purpose of this is to allow pluggable elements to hook into the
widget rendering process and to intercept widget rendering in edit mode
during certain conditions. These conditions are specific to the
registered component.