Change `WidgetOverrideElements` from `throw` to explicit function call.
Prior to this change, components passed to `WidgetOverrideElements` were
expected to throw a component when overriding the widget's edit
component was required. This lead to an unwanted error being logged to
the console, even if the thrown component was caught.

This change replaces the usage of `throw` with an explicit function call
that is being passed to the component when rendered. The component can
call it conditionally to override the widget's edit component.