Add pluggable elements rendered in widget edit mode. (#6992)
* Renaming `error` field to `errors` for consistency.

* Return undeclared parameters from search type query strings as well.

* Placing query elements in correct position.

* Supplying if editing query on query tab or widget edit mode via context.

* Supporting widget override elements.

Widget override elements can be registered and are rendered when editing
a widget. These component can either return elements which are rendered
next to the widget, or they can throw a renderable component which will
be rendered instead of the component.

The purpose of this is to allow pluggable elements to hook into the
widget rendering process and to intercept widget rendering in edit mode
during certain conditions. These conditions are specific to the
registered component.

* Passing function to reset override state.

* Allowing to pass backdrop to modal.

* Propagate actual errors, add tests.

* Use `retry` for consistency.

* Adapting test fixture.

* Removing unused import.

* Extracting `suppressConsole` test helper.

* Change `WidgetOverrideElements` from `throw` to explicit function call.

Prior to this change, components passed to `WidgetOverrideElements` were
expected to throw a component when overriding the widget's edit
component was required. This lead to an unwanted error being logged to
the console, even if the thrown component was caught.

This change replaces the usage of `throw` with an explicit function call
that is being passed to the component when rendered. The component can
call it conditionally to override the widget's edit component.

* Adding missing semicolon.

* Removing unused `suppressConsole`, wrapping test which requires it with it.

* Adding missing semicolon.