Change `Show documents for value` to use widget query instead of filter
With the introduction of separate queries per widget, widget filters are
becoming deprecated. They are usable from the API, but there are no
features in the UI enabling to use them. Therefore, this PR changes the
`Show documents for value` action to use widget queries instead of
widget filters. In addition, it creates a helpful title for the newly
created message table.

Fixes #7020.