Ignore case when deserializing dashboard widgets in migration.
Prior to this change, the case of a widget's type field must be in upper
case for its config subclass to be resolved properly. Content packs use
lowercase for these when installing a dashboard, so these dashboards are
not migrated properly.

This PR is adding a custom type resolver for deserializing widget config
subclasses, which is uppercasing type ids before resolving them.

Fixes #7035.