Allow queries with empty search types.
This change lifts a restriction where the backend raises an exception
when a search contains a query with an empty set of search types.
Searches like this can be generated for:

1.) new dashboards containing no widgets
2.) dashboards containing only widgets which do not generate search
types (e.g. unknown 3rd party widgets migrated over)

In the first case the resulting error message is only briefly shown to
the user, in the second case all widgets contain this error instead of
showing the "Unknown widget" help text.