Register search type generator for default widget.
This change is registering a search type generator for default widgets
(which are also used for migrated unknown 3rd party widgets), which just
returns an empty list.

This is necessary, because otherwise regenerating a search for a
search/dashboard containing an unknown widget fails, as no function
generating search types for this widget can be resolved.