Destructure result from `SystemStore` properly in `Footer`. (#7077)
When refactoring the `Footer` component, it was overlooked that the
state of the `SystemStore` is wrapped in an object with another `system`
key and tried to access its contents directly. Thus the footer did not
show the Graylog version and the current hostname.

This change is now adding a props mapper function to the `connect` call
which destructures the store's state properly and adds tests.