Reducing remounting/-rendering of `MessageDetail` component. (#7085)
Prior to this change, the `NodesStore` triggered a periodic rerendering
of the `MessageDetail` component. This caused a remounting of the
`FieldActions`/`ValueActions` component, resetting the state of
potentially opened drop downs.

This change is decoupling the consumption of the `NodesStore` by
reusing the `LinkToNode` component, keeping the store consumption as
local to rendering as possible.

Fixes #7059.