Backport: Add URL whitelist service (#6854)
* Initial support for reading and writing url whitelist via REST API

* Add permissions for reading and writing url whitelist

* Fire audit event for url whitelist change

* Implement matching strings against whitelist

* Add shortcut method for whitelist check

* Bind url whitelist service as singleton

* Use whitelist service in JSON path data adapter

* Use whitelist service in DSV HTTP data adapter

* Use whitelist service for HTTP event notifications

* Improve error propagation

When the URL is not whitelisted, we want that error to be exposed so
that it can be displayed on the frontend. When the url becomes
whitelisted, we want the error to go away and any previous or new errors
related to parsing the DSV file or requesting the file should be
exposed.

To make this work reliably, we have to make sure that in case of a
whitelist error the file is fetched again so that the errors related
to file fetching and/or processing surface again. Also we want errors
related to parsing the file show up in the UI, which was not the case
before when the file became corrupted at the source when the data
adapter was already running. Those errors would only show up when the
data adapter was (re-)started.

* Provide REST endpoint for checking URLs against the whitelist

* Add @NoAuditEvent annotation to validation endpoint

* Check URL whitelist when validating data adapters

* Add license headers

* Add migration to create initial whitelist

* Add missing jackson annotations

* Add system notification on whitelist failure

Set a system notification when a url doesn't pass the whitelist check
during operation.

* Indicate whitelist problem with JSONPath adapter by setting error

Use the same mechanism as with the DSV over HTTP adapter to indicate
that a whitelist error happened.

* Add "id" and "title" fields to whitelist entries.

* Add ability to disable a whitelist

* Add basic content pack support for url whitelist entries

* Add simple caching of whitelist in whitelist service

Simply try not to fetch the whitelist from mongo DB each time it is
requested. May fetch the whitelist from the DB more often than it needs
to but we'll accept that in favor of adding synchronization.

* Add REST endpoint to validate java regular expressions

* Add service to conveniently publish url whitelist failure notifications

* Use annotations in rest resource for permission checks and validation

* Add whitelist checks and migration for URL in legacy http alarm callback

Did not add a system notification for runtime whitelist failures of this
 kind as this wouldn't be straight forward. This is legacy code so we
 should be ok without this.

* Use AutoValue classes for whitelist entries. Make "title" required

* Clear cached whitelist on write.

Cluster events don't propagate fast enough, so we have a race condition
when saving subsequently on the same node. Mitigate that by clearing the
 cache explicitly.

References #6853

(cherry picked from commit f0c9374b9f32670c354857715ffbc08d1a65d146)

Conflicts:
	graylog2-server/src/main/java/org/graylog2/contentpacks/ContentPacksModule.java
	graylog2-server/src/main/java/org/graylog2/contentpacks/model/ModelTypes.java
	graylog2-server/src/main/java/org/graylog2/lookup/LookupModule.java
	graylog2-server/src/main/java/org/graylog2/lookup/adapters/HTTPJSONPathDataAdapter.java
	graylog2-server/src/main/java/org/graylog2/migrations/MigrationsModule.java
	graylog2-server/src/main/java/org/graylog2/rest/resources/system/lookup/LookupTableResource.java