Duplicate query/timerange/streams/filter for specific field/value actions.
Prior to this change, several field actions ("Chart", "Aggregate",
"Statistics") and the "Show documents for value" value action did not
duplicate the context of the widget they were triggered from, although
the user most certainly expects that. This led to differing results.

When a user e.g. clicks on the "Show documents for value" action on a
value of a filtered down (or extended through a bigger time range)
result, the widget created from the action did not contain the same
query/timerange/streams/filter settings but the default ones, and
therefore referred to a different set of documents.

This change is creating a shared function which duplicates these
settings from a given widget and is supposed to be used by any action
that creates a new widget that is supposed to work in the same context
as the widget it was triggered from. This shared function is used in the
aforementioned actions.