Disabling field actions if not applicable to current field/function.
This PR is correcting the conditions for enabling `Aggregate`,
`Statistics`, `Add to all tables` & `Remove from all tables` if the
current field is a function.

In addition, it changes the `Chart` action to use the function itself as
the series for the newly created widget if it was triggered by clicking
on a function instead of a field.

Fixes #7089.