Enforce system wide TLS protocol settings
With the introduction of netty 4 in Graylog 3.0, we changed the default
SslProvider to OpenSSL.
While this gives a better performance, it lacks the ability to disable
TLS protocols that are considered flawed.
Changes to the java security properties are only respected by jdk based
SslProviders (jersey / grizzly in our case).

Introduce a system wide default setting for supported TLS protocols,
with a sane default to TLSv1.2 (and TLSv1.3 if the jdk supports it).
These protocol settings are applied to netty based inputs and
to the webserver.

Changes can be made with a new server setting:
`#enabled_tls_protocols = TLSv1.2,TLSv1.3`

Fixes #6444