Convert week/month bucket interval to days if value is bigger than 1.
Due to limitations in Elasticsearch, it does not support time unit
bucket intervals with values bigger than 1 for units bigger than days.
Therefore, this PR transparently converts these to days (with multiples
of 7 resp. 30) if the value is bigger than 1, leaving the original unit
otherwise.

In contrast to the initial attempt in #7055, this PR is performing the
change in the backend, transparent to the caller and relieving the
caller from having to do the conversion before.

Fixes #7022.