Add current function to value path. (#7108)
* Return original term if new term is empty.

* Adjust flow types for helper function.

* Extract field from value instead of key if value path segment is `_exists_` clause.

* Fix conditional rendering of `Value`/`EmptyValue`.

Fixes #7104.

* Add current function to value path.

Prior to this change, only field:value pairs were added to the value
path of a value in a data table. This led to empty/incomplete queries
being generated from the "Show documents for value" action.

This change is now adding functions to the value path as well if they
refer to a field. A function like this generates a `_exists_:<field>`
clause, so all documents containing this field match.

Fixes #7103.
Fixes #7104.

* Correcting conditional.

* Do not include source/timestamp fields twice in resulting widget.