Avoid using PathConfiguration#getDataDir() for default path configs
Using "#getDataDir()" to initialize the default dir values for the
contentpack and message journal works, but it implies that they
will automatically use the user configured "data_dir" setting. This is
not true because "#getDataDir()" to set the default value will be called
before the configuration is parsed.

Introduce constants for the bin, data and plugin dir and use those as
default values to make it explicit that the default will not change
automatically by setting a custom "data_dir".