Fixing effective time range in pivot search type results. (#7133)
Prior to this change, the effective time range of the pivot was used for
the search type's result only if the result contains no documents
(causing the min/max aggregations to return `null`). This led to the
effective time range for the pivot search type being the
smallest/largest timestamp of its documents, instead of the
beginning/end of the supplied time range.

This PR is now changing this behavior to do one of these:

  - If the effective time range of the search type is set to "All
messages" and the search returns documents, the effective time range of
the result is the smallest/largest timestamp of the result set
  - If the effective time range of the search type is set to "All
messages" and the search returns no documents, the effective time range
of the result is set to unix epoch 0/now
  - In all other cases, the time range of the result is set to the
effective time range of the search type

Fixes #7128.