Synchronize timerange & query with URL on search page.
The current search has a feature, where the current time range and query
string is synchronized with the URL in the browser. This means that:

  - current time range and query string are part of the URL and copying
and replaying the URL allows restoring the state for these
  - a change to time range or query string creates a new state in the
history stack, allowing the user to go back and forward between
different states
  - going back or forward in the browser history restores the time range
& query string of this specific state

This PR is recreating this functionality for the new search.

Fixes #7144.