Add mapping of stream ids in view content packs
Prior this change, when creating a content pack containing
views with streams the stream was collected as an dependency
into the content pack but the StreamEntity got a new id.
This id needs to be used in all references of the stream.
Therefore also in the ViewEntity where a streamid is persisted.
Also when installing this content pack the installed
stream has a new ID yet again and the view needs to know
this as well.

This change adds this mapping to the Widget and WidgetEntit
for the DashboardWidget.
For the search types there are now Entities so we can easily
map the streams during creation and installation of the
content pack.
For the query we needed to add a generic Builder access
for the Filter so we could change the Stream id for
a StreamFilter and rebuild the Filter refered in the
QueryEntity and Query.