Add support to dump the current ProcessBuffer state
Introduce a new API call that returns the messages
that are currently processed in each ProcessBufferProcessor.

This is useful to debug cases where message processing got stuck.
The most common case is the use of an inefficient regular expression
in extractors or pipeline rules.
Understanding which regular expression caused the problem can be
difficult, without knowing which message triggered the problem.

Example:

- Set up pipeline rule with an evil backtracking regex like `/(x+x+)+y/`
- Send message that spends hours in processing.

- GET https://172.16.1.1:9000/api/system/processbufferdump

```
{
  "processbuffer_dump": {
    "ProcessBufferProcessor #4": "idle",
    "ProcessBufferProcessor #2": "MessageEvent{raw=null, message=source: t480 | message: xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx { application_name: mpf | level: 5 | gl2_remote_ip: 127.0.0.1 | timestamp: 2020-01-16T11:00:48.296+01:00 } [...] }",
    "ProcessBufferProcessor #3": "idle",
    "ProcessBufferProcessor #0": "idle",
    "ProcessBufferProcessor #1": "idle"
  }
}
```