Do not parse empty value of from/to field as date for absolute ranges.
Before this change, emptying one of the date fields of the absolute time
range inputs resulted in an error being logged to the browser console,
because it was tried to parse as date.

This PR is checking if the supplied new value is a valid date string
before parsing it.

Fixes #7190.