Reset time in SearchesCleanUpJobWithDBServicesTest teardown method
This adds a missing DateTimeUtils.setCurrentMillisSystem() to avoid
a sporadic test failure in DBJobTriggerServiceTest.

The SearchesCleanUpJobWithDBServicesTest is using
DateTimeUtils.setCurrentMillisFixed() to set the time to a specific
value. This requires the usage of DateTimeUtils.setCurrentMillisSystem()
to reset the time to the current time. Otherwise tests running in the
same JVM using the current time might fail.

I noticed that the DBJobTriggerServiceTest#deleteCompleted test method
failed from time to time on a new Jenkins worker machine. This is
probably due to a changed execution order because the machine is faster
than the older worker machine. Still, strange that it didn't show up
earlier.