Add validation to multiple aggregation condition (#7219)
* Fix aggregation summary without expression

Check if no condition was configured and show a message accordingly.

* Clear conditions when last condition is removed

This allows users to remove the aggregation from a Filter and
Aggregation Event Definition.

* Do not show condition validations inline

The server currently cannot provide accurate inline validation, so we
display any server validation errors above the conditions section.

* Validate aggregation conditions as they change

To avoid annoying users with errors as they type, display a global
status next to ghe condition summary, and a button to toggle inline
errors.

* Add conditoin validation in summary page

* Display error when invalid aggregation condition

If submitting an invalid aggregation condition returns a 400 response,
display a somewhat nicer and persistent error.

* Catch more class names appearing in 400 responses

* Also check if the function is really set

* Update snapshots

* Unify event definition validation

Use server response to validate undeclared parameters. This way we have
less plugin-specific code in the common path, and it works the same way
as other validation.