Fix GlobalOverride behavior (#7250)
* Fix GlobalOverride behavior

Prior to this change, the globaloverride for timerange was overriden by
the query builder for globaloverride. That way a globaloverride for
query and timerange did not apply the timerange of the override.

Also when disabling the timerange override in the gui the store did
set a plain javascript object instead of a GlobalOverride value class,
with that the Store could not change the timerange of the global
override anymore.

This change will use a common Globaloverride Builder for `hasTimerange`
and `hasQuery` that way the query does not override timerange.

Also when disabling the timerange in the GUI we now only unset the
timerange value of the Globaloverride.