Change data structure in WidgetFormattingSetting
Prior to this change, the data structure to store
WidgetFormattingSettings did use the "user input" for storing a map.
But mongodb does not allow to store keys with dot in the map.

This change will change the data structure from a map to a list
of objects. That way mongodb does not need to store the keys.