Conditionally set "gl2_message_id" and use message timestamp
Previously the timestamp part of the "gl2_message_id" was the processing
time. Since we have the actual message event time after processing, we
want to use that one.

We also now check if the field has been set already. If that's the case,
we don't overwrite it. (forwarder use case)