Reload search/dashboard again if id props changes.
Before this change, loading multiple searches consecutively, exporting a
search to a dashboard or saving a dashboard as a new one and navigating
back and forth in the browser history would not trigger reloading the
corresponding searches/dashboards.

This change is now implementing the required check which triggers
reloading a search/dashboard if the id of the current search/dashboard
in the URL changes.

Fixes #7298.
Fixes #7299.