Fix default selection for stream output assignment
When loading the stream output page, the dropdown for "Assign existing
Output" shows the first existing output as selected. When pressing the
button, an error had been generated because there wasn't an actual
selection.

This change is fixing the AssignOutputDropdown to ensure the PLACEHOLDER
variable is initialized before it is used.