Change color chart from Map to List of Value Class  (#7306)
* Change data structure in WidgetFormattingSetting

Prior to this change, the data structure to store
WidgetFormattingSettings did use the "user input" for storing a map.
But mongodb does not allow to store keys with dot in the map.

This change will change the data structure from a map to a list
of objects. That way mongodb does not need to store the keys.

* Migrate existing data structures for color charts

Prior to this change, we changed the data structure of color charts
from a generic Map to auto value classes. This change prevents
existing installations to run.

This change will migrate the existing datastructures in the mongodb
to the new auto value classes.

* Add missing licenses

* Fix annotations from @dennisoelkers

* Add more tests