Fix blocking issue in DNS lookup data adapter
We cannot use "Future#sync()" to wait for the result of the DNS
resolver. This can block forever (until the server gets restarted)
when the DnsNameResolver and the netty event loop gets shut down while a
DNS request is running.

This happened on busy systems that do a lot of DNS lookups and the DNS
lookup data adapter gets restarted while DNS requests are running. (e.g.
by updating the data adapter's configuration)

It can get even worse when multiple processing threads try to lookup the
same hostname or IP address at the same time, using the same lookup
table. One thread is waiting for a response but never returns.
All other processing threads are blocking on the cache lock, waiting on
the other thread to return.

The solution is to not use "Future#sync()" and use "Future#get()" with a
timeout so the thread doesn't wait for the future completion forever.

By default the timeout for waiting on the future is based on the query
timeout plus 100 ms to avoid canceling the future before a potential
DNS query timeout kicks in.

This change also ensures that the DnsNameResolver will be closed before
shutting down the event loop.

Fixes #5782