Allow *Create Extractor* for all types (#7335)
* Allow *Create Extractor* for all types

Prior to this change, the `create extracotr` field action was only
enabled for string values. This lead to the problem that field
from type `text` and `string` (compound) could not be extracted even
though it would work perfectly fine.

This change will remove the check and allows extrator creation for all
field types.

Tested with: string, number and timestamp.

* Fix test for value action