Increasing timeout for element finders in `CreateNewDashboard` test. (#7381)
When run in parallel, the `CreateNewDashboard` test has turned out to be
flaky. A further attempt to increase the jest timeout for the test suite
did not fundamentally improve the situation.

This change is increasing the timeout for the individual element finders
in the test from the default of 4500ms to 15000ms, hoping to make the
test more reliable.