Fix duplicating of query tab if no widget has been renamed. (#7397) (#7414)
* Fix duplicating of query tab if no widget has been renamed.

Before this change, duplicating a query tab containing no widgets or
where no widget had been renamed or given a title before, would fail.
This happened because during query tab duplication the map of widget
titles was extracted from a nested titles structure, which - at this
point - does not contain a widget titles map.

In addition, a second error occurs after successfully duplicating a
query tab: the `QueryTitle` component tries to switch to the new query
tab by using the id of the new query supplied by the promise returned by
the duplication action. Unfortunately the duplication action does not
return the query id. Nonetheless, the new query tab is selected, because
of the `ViewStore#createQuery` action being called.

This change is doing two things:a

  * Fall back to an empty map if the widget titles map is not present at
this time.
  * Remove the explicit call to switch to the newly created query tab.

* Fixing linter hints.

* Adapting test to changed behavior.

* Removing unused import.

Co-authored-by: Dennis Oelkers <dennis@graylog.com>