Handle message fields types not being available (on time).
Prior to this PR, several users reported that under some circumstances
the message list widget raises an exception. Following the stack trace,
it looks like field types might not be available (being `undefined`) and
the `find` method called on it nevertheless, resulting in an exception.
This might be because of either a race condition or a missing/empty API
response.

I was unable to reproduce the error in my development setup while using
the product, even not by delaying the fields types store's response.
Nevertheless, I was able to reproduce it in the added test case, by
deliberately providing an `undefined` `fields` prop, allowing me to fix
the related code lines.

Fixes #7365.