Allow ordering of configuration fields
Fields in generated configuration forms are not placed in any specific
order. (Apart from the fact that optional fields are ordered at the end)

With this change it is possible to specify an explicit position
attribute on a config field.
The default positioning value is 100.
If you want your field higher up in the form, choose a value below 100.
If you want it at the end, set it higher than 100.