Fix call of createNewVersionFromRev on undefined (#7442)
Prior to this change, createNewVersionFromRev was called on undefined
because contentPackRevisions was assumed to be in the state after
ContentPacksActions.get was called, but it was not.

This change will use the parameter from then action of the get function
which contains the contentPackRevisions.

Fixes #6690

(cherry picked from commit 650c729fa2506da95eadcb7e1a7e71530ce02b30)