Fix ClassCastException in MongoDbSession#getTimeout()
Cast the value to a Number instead of a Long and then use ".longValue()"
to avoid the exception.

This mostly happened when the session timeout was -1 and MongoDB
returned an integer instead of a long for that.

It has been broken for a long time but it never showed up because the
ExecutorServiceSessionValidationScheduler in shiro didn't catch
exceptions and had no uncaught exception handler.

Since shiro 1.5.0 this has been fixed and we see the ClassCastException
in the graylog logs.