Don't expose KafkaJournalConfiguration via the HTTP API
This fixes the "/api/cluster/<node-id>/journal" API endpoint to not
throw an error anymore.

In #7359 we introduced the PathConfiguration#getNativeLibDir() method
and this broke deserialization of the JournalSummaryResponse because
"#getNativeLibDir" was automatically serialized but couldn't be
deserialized.

This change introduces a new KafkaJournalConfigurationSummary value that
we can serialize and deserialize in the API. Using this also avoids
exposing new fields from Pathconfiguration or KafkaJournalConfiguration
automatically.
This also removes the constructor and jackson annotations in
KafkaJournalConfiguration beacuse they are not needed anymore.

Fixes #7521