Use AES SIV to encrypt the access tokens
This ensures deterministic cipher text values so we can use the value to
lookup the tokes in the database.

We need this because our API token implementation only knows about the
token and has no user account or token name information. Changing this
to get a more secure implementation would be a breaking API change.