Specify ES search timeout to match the HTTP request timeout
The `elasticsearch_request_timeout` configuration value will set the
HTTP request timeout of ES operations but it does not tell ES to cancel
the search request with that timeout. This means the request will
continue to run in ES to completion while the client has timed out. In
turn, this could cause the user to resubmit the request which would
issue another long running query to stack against and overwhelm the
ES cluster.

Fixes #4614