Support legacy (pre 3.2) search result in data table visualization. (#7611)
Before `3.2`, search types did not had a `name` attribute. In 5ee5115 an
optional name attribute defined by the caller was introduced. Views
created before this, did not contain it and search results are scoped
with the search key id instead. This is handled by all visualizations
but the data table.

This leads to views containing data tables not rendering properly in
`3.2`, if they had been created before `3.2`, as reported in #7565.

This change is now adding support and tests for this for the data table.

Fixes #7565.

Co-authored-by: Dennis Oelkers <dennis@graylog.com>