Registering callbacks even if search refresh condition hooks fail.
Before this change, the `ExtendedSearchPage` component did not register
callbacks for `ViewActions.search.completed` & `SearchActions.refresh`
if the search refresh condition returned `false` on the initial attempt.

The impact of this is that e.g. when the enterprise plugin is installed
and a search page is opened with an initial query that has unregistered
parameters, the callbacks are never registered and therefore from this
point future updates to the search or refresh requests are not handled.

This change is registering callbacks in any case, regardless of the
search refresh condition results.