Keep DataTable rows up-to-date
To fix issue #5965, we moved some code calculating the effective rows
into `componentDidUpdate`. This unfortunately has the effect that when
we receive new rows in the props, we may try to render an unexisting
row once, before updating the list that is kept in the component's
state, causing #7544.

To avoid that situation, this commit will avoid using `filteredRows`
when the `DataTable` doesn't have any `filterKeys`, that is, when we
don't display a filter. In summary:

- `DataTable` without filter: use rows coming from the props instead of
  syncing them in the component state
- `DataTable` with filter: filter takes care of updating the
  `filteredRows` state key, but `DataTable` filters out any keys that are
  not in `rows`

Fixes #7544