Add support for high level Kafka consumer mode (#7504)
* Add support for newer Kafka versions

Newer Kafka versions are supported by using the new high level consumer API.
Lucky for us, support for the new API is already included in our currently used Kafka
client library. This means we can avoid updating it, and still be
compatible with our on-disk journal.

Instead of creating a new Transport class, use a legacy switch to keep
supporting older Kafka versions.

Most of this was spearheaded by Muralidhar Basani in
 https://github.com/Graylog2/graylog2-server/pull/4770
Thank you very much! :-)

Co-authored-by: Muralidhar Basani <murali.basani@gmail.com>

* Drop stale comment

* Allow ordering of configuration fields

Fields in generated configuration forms are not placed in any specific
order. (Apart from the fact that optional fields are ordered at the end)

With this change it is possible to specify an explicit position
attribute on a config field.
The default positioning value is 100.
If you want your field higher up in the form, choose a value below 100.
If you want it at the end, set it higher than 100.

* Fix validation, adjust comments

* Add support for custom properties

This allows users to configure things like TLS.

* Set system wide default for TLS protocols

* Revert "Set system wide default for TLS protocols"

Better leave this entirely up to the user

This reverts commit dabaa00d4f3891f31c8a56c8a3194b8ecd4f7507.

* Avoid logging ConsumerConfig

It might contain sensitive credentials

* Adjust comment

* Improve Exception handling and refactor code

Avoid resource leaks by handling more exceptions.
Shutdown the input on unrecoverable exceptions.
Retry on others.

Add workaround for hanging poll if kafka is down:
 https://issues.apache.org/jira/browse/KAFKA-4189

* Restart Input after unrecoverable Kafka error

Also fix thread leak. We need to shutdown the ExecutorService.

* Don't restart the Input after a KafkaException

This reacharound via the eventbus is too ugly
for a problem that I can't trigger in practice.

Simply log the exception and stop the transport.

* Stop transport if we can't process the records

If our processing throws an exception while
consuming the records into our journal,
it's better to stop the transport early so we
don't lose too many messages.
(enable.auto.commit is our default)

* Simplify consumeRecords()

With the old kafka client, every next() call would mark the offset
as commited. With the new client, this isn't the case anymore,
thus we can simplify the code.

* Improve position defaults

Co-authored-by: Muralidhar Basani <murali.basani@gmail.com>
Co-authored-by: Bernd Ahlers <bernd@graylog.com>