Extract IndexLookup from ElasticsearchBackend
Previously the lookup of indices for streams and time range was done
in ElasticsearchBackend by loading indices, streams, and then finding
matches. It was impossible to reuse this logic elsewhere.
Furthermore this made ElasticsearchBackend harder to test, because
a lot of setup was required to make sure that proper inddex ranges and
streams are loaded during the test.

This commit:
- Extracts that logic into a separate class `IndexLookup`
- Simplifies tests for ElasticsearchBackend accordingly
- Adds tests for `IndexLookup`