Raise MissingStreamPermissionException for having no streams at all
Prior to this change, if a user did not have permissions to any streams
we raised a exception which would lead to an 404 error on the
frontend.

This change will raise an MissingStreamPermissionException which
will mapped to a NOT_ACCETPABLE (406) error code which will be
treated differently on the frontend side and not show a not found error.