Add search support for custom DateRange aggregations
This extends the search backend to support Range Aggregations with Dates.

 https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-range-aggregation.html

Before we were using a DateHistogram which was creating problems because
the buckets were rounded to the nearest full minute / hour, depending on
the range.

The PivotAggregationSearch can now wrap each search with precise range buckets,
which gives us the results we need.