Prevent default action for form submit event in `PivotConfiguration`.
Before this change, the default action for the form submit event
triggered by submitting the `PivotConfiguration` form was not prevented.
This lead to a redirect happening in some browsers (namely Safari),
while working as expected on others (Chrome/ium).

This change is now preventing the form submit event to fix this for
Safari and others.

Fixes #7806.