Migrating the InputStatusService from the enterprise plugins repo to the main repo.
The InputStatusService was created to support Inputs that need to persist state for the duration of their lifecycle.  It was created to allow the Okta input to checkpoint on seen messages in order to avoid duplicate message processing when the input is restarted.

Original PR: https://github.com/Graylog2/graylog-plugin-enterprise-integrations/issues/136