Add stages to result of parse pipeline api endpoint result
Prior to this change, the stages setter was not called for
the PipelineSource builder for the result of pipeline
parse api endpoint. This led to an 500 server error
and the api endpoint could not be used.

This change will take the parsed stages from the pipeline
and add them to the resulting PipelineSource.

Fixes #7322